Abstract factories and parsers (#5)
Modifies existing parsing code to take a subclass of BirthRecord and an instance of RecordFormat as parameters and uses these to setup the parser to read any files. Makes RecordFactory an abstract class, which is chosen based on the values provided in parameters.
* Implemented abstract factories and parsers
     * RecordParser now takes a type T and generates Records of class type T
using an implementation of the IRecordFactory interface
* Added/updated comments
* Introduced RecordFormat support
     * RecordParser now takes a format when it is instantiated and uses this to
parse records. Also replaced IRecordFactory interface with RecordFactory
abstract class